+++ gammapy/shower/tests/test_hillas.py
@@ -1,7 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-from __future__ import print_function, division
-from ...shower import hillas_parameters
-
-
-def test_hillas_parameters():
-    pass